requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
guppy3==3.0.9
idna==2.8
PyMySQL==0.9.3
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
